ported slave functionality to plugin system, added controlsocket plugin
Not much new code, mostly movement to the slave plugin. Added a controlsocket
plugin that doesn't do a whole lot yet, just listens for a signal to quit on
the socket. Will flesh it out more later, but I need that functionality to kill
daemon's for testing. qsrvutils.py no longer needed since all of it's code was
moved to master and slave plugins.